+++ lib/cart_screen_widget/Menu.dart
@@ -1,0 +1,85 @@
+import 'package:flutter/material.dart';
+import 'package:intl/intl.dart';
+
+var f = NumberFormat('###,###,###,###원'); // 천단위 구분기호 추가
+
+class Menu extends StatelessWidget {
+  const Menu({Key? key, this.menuName}) : super(key: key);
+  final menuName;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.white,
+      child: Column(
+        children: [
+          Row(
+            children: [
+              SizedBox(
+                width: 20,
+              ),
+              Text(
+                menuName,
+                style: TextStyle(
+                  fontSize: 17,
+                  fontWeight: FontWeight.bold,
+                ),
+              ),
+              Spacer(),
+              IconButton(
+                icon: Icon(
+                  Icons.close,
+                  color: Colors.grey,
+                ),
+                onPressed: () {},
+              ),
+            ],
+          ),
+          Row(
+            children: [
+              SizedBox(
+                width: 20,
+              ),
+              Container(
+                decoration: BoxDecoration(
+                  border: Border.all(
+                    color: Colors.grey.withOpacity(0.3),
+                    width: 1,
+                  ),
+                  borderRadius: BorderRadius.circular(12),
+                ),
+                child: ClipRRect(
+                  borderRadius: BorderRadius.circular(12),
+                  child: Image.asset(
+                    'images/chicken.png',
+                    width: 70,
+                    height: 70,
+                    fit: BoxFit.cover,
+                  ),
+                ),
+              ),
+              SizedBox(
+                width: 10,
+              ),
+              Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    '• 찜 & 리뷰 약속 : 참여. 서비스음료제공',
+                    style: TextStyle(
+                      color: Color.fromRGBO(125, 125, 125, 1.0),
+                    ),
+                  ),
+                  Text(f.format(18000)),
+                ],
+              ),
+            ],
+          ),
+          SizedBox(
+            height: 20,
+          ),
+        ],
+      ),
+    );
+  }
+}
