+++ lib/cart_screen_widget/StoreName.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class StoreName extends StatelessWidget {
+  const StoreName({Key? key, this.storeName}) : super(key: key);
+  final storeName;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.white,
+      height: 70,
+      child: Row(
+        children: [
+          SizedBox(
+            width: 20,
+          ),
+          ClipRRect(
+            borderRadius: BorderRadius.circular(12),
+            child: Image.asset(
+              'images/chickenCartoonImage.jpg',
+              width: 35,
+              height: 35,
+            ),
+          ),
+          SizedBox(
+            width: 10,
+          ),
+          Text(
+            storeName,
+            style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
+          )
+        ],
+      ),
+    );
+  }
+}
